Revert "Add /bootstrap-apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92af3f29-73d8-4290-b5b2-e4616a053193). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id28b12e0cadf0229bc03b8a1a5bcd740e46de1b8