Revert "Add /bootstrap-apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd65b239-0945-486e-8404-2163c232d175). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I708a54e0a2cd7a81f4793f90b815d950f25a2f67