Revert "Add /bootstrap-apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1616589c-b0fb-4fe5-8610-076e2f368028). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9aca3a90f631180bf30ddd675f4c9436537b233d