Revert "Allow vold_prepare_subdirs to use apex_service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e69f6690-745e-479d-ac45-f4be88d76264). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I654de6320167ca7699b6a4ea6e0ff9845197e955