Revert "Allow vold_prepare_subdirs to use apex_service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1153f00e-a01e-477d-9119-520fc3ea9863). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I84b2251961968a5942f56698f34d57ac08a216c1