Revert "Allow vold_prepare_subdirs to use apex_service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb08b06b-016d-4052-b703-0d40e9c781ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie58542cae331d2462a86de8e6170dbfdc451bfe1