Revert ntfs file context changes
Technically, this could break other potential implementations of NTFS in
future. I am not rolling back the adjoining fuseblkd_exec and fuseblkd_untrusted_exec
code, because secure implementations of fuseblk drivers should still endeavour
to use this more secure implementation.

However, we should give implementors the choices whether to use it or
not, even if it will open the door to potentially less secure
implementations.

Bug: 294925212
Test: Builds and boot.
Change-Id: I6d3858517e797b3f7388f9d3f18dd4a11770d5bc