Revert "Make coredomain violation as a build error"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f40f2f9d-b271-41e0-a7af-c56d37959552). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ee13bf4df34dd14bb6b5263be874f0f2f61be52