Revert "Make coredomain violation as a build error"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14562644-65e3-48d7-8ec5-7c355dc77d1c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifc274c04736742ab1840681af662162b236f83ac