Revert "C2 AIDL sepolicy update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eaa9cdb8-f56d-4723-987b-f0a84afa1ba8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica1568126881fd1a8b9fcc6c5ecf4fc8f2c79c51