Revert "Add a fuzzer for media.c2 HAL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd94fae3-7415-40dd-b0d3-612906079279). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id81688b5fcab10ec84f045e40a0626d2545c5b90