Revert "SELinux policy changes for uprobe."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49d7bb6c-b73c-46e3-a757-3b9b06da7264). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a74b77bafe7987c8c6ada284ecbd9f94ee37d36