Revert "SELinux policy changes for uprobe."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/603a280b-b770-4441-a353-4d65165b0010). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6012f6ffea539c6e495cfa228ea581cd3f646f08