Revert "sepolicy: Rename NFC apex package name"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f893ea8e-a6c2-4b4d-ac06-1192a75b8424). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic178a02389d0ba98f9ca83f6b5760a9a762a8335