Revert "sepolicy: Rename NFC apex package name"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13b75b8a-363e-4ca3-a9e9-5e3c37d04a65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ba771e3de426c892d53f4c893ef8324613bcc72