surfaceflinger: Add perms to write to "/data/misc/sfbuffers/"
A new debugging feature is being added to SurfaceFlinger that allows
developers to dump cached buffers as PNGs. This feature requires a place
to write the buffers, which will be in the new folder,
`/data/misc/sfbuffers`.

Because of existing SELinux rules, we need to explicitly allow-list this
folder for writing by the SurfaceFlinger binary.

Bug: 306193257
Bug: 309581086
Bug: 307862724
Test: Ran the backdoor locally, successfully
Change-Id: I1b878dbb4639b144ad026b9c577c5436e2dc5db5