Allow system_server to read multicast routing packet counts
Multicast forwarding cache (MFC) is added to support multicast
routing. ioctl with SIOCGETSGCNT_IN6 reads the packet counts
for an MFC, this can be used to determin if an MFC has been
actively used or can be recycled if the packet count has not
changed for a certain amount of time.

Test: Verified on cuttlefish that system_server has the permission
to call ioctl with SIOCGETSGCNT_IN6 on a raw ip socket

Change-Id: Id9c8ed45a4cfb032655a78daf1f024ee0d34d594