Revert "Added entries for audioflinger and audiopolicy aidl fuzzer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f0d2c8ff-5175-4fe9-a62d-f954d8ede0d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I605ec066b409fba73beeee46ca99994d2f4ed7fb