Revert "Add permission for VFIO device binding"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d0ef49f-7d01-46c5-a1e3-5bf02be33bbc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I583277de4cdf98899dfcf4354abe088cca2b1bd0