Revert "Allow system_server to communicate with virtual_camera"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5768a9c7-80af-4b7a-aea9-86813ce80054). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic5d941504fd396206fff41e1d96b15660b24b166