Revert "Allow system_server access to hidraw devices."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2540f4fb-d944-45bf-9106-72c818b95ead). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e07bac8ee663ba20d657c7920a594873c76336c