Revert "Allow remount to update the super partition."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a605e2d-9765-4c52-b8a7-5a20594b972d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib597438fcec333e7878eed328eeff199269e103c