Revert "Allow CAP_SYS_NICE for crosvm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bef269d8-679f-41e6-af5e-0d6d7ee17c1f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I182b14585daf3fe9c32a2bd724613a7a3288bde1