Revert "Fix freeze test condition to board api"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9414a9f2-6745-40d6-9690-c0b8373bb4cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6051b7fb5b40169003527d8a783bce255a0b3d1d