Revert "Fix freeze test condition to board api"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0bce28b4-ddda-4b52-8602-9136874aa1e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idbd2d100a5150d7e054b767871d231e9ea4501d4