Revert "Add 1000000.0 mapping file temporarily"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0bce28b4-ddda-4b52-8602-9136874aa1e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id6cb14b96d83cd1af1b45ef23787ff9822f057aa