selinux setup for files under /metadata/aconfig dir
Initially set it up to be accessible by system_server process. When
aconfig storage daemon is created, will add another file permission to
storage daemon.

Context to why we are hosting flag data on /metadata partition:

Android is adopting trunk stable workflow, flagging and A/B testing is
essential to every platform component. We need some place to host the
flag that are accessible to system processes that starts before /data
partition becomes available.

In addition, there has been a long discussion regarding utilizing /metadata partition for some process data, another example is mainline modules, we are trying to make them to be able to be mounted earlier, but cannot due to /data availability.

Bug: 312444587
Test: m
Change-Id: I7e7dae5cf8c4268d71229c770af31b5e9f071428