Revert "Relax neverallows for vendor to use /system/bin/sh"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef0ed6dc-6c99-4275-8510-0304fbdd06f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1fd5a3385b24a5b401d2a17334c21f106dfd277