Mark libft2.so installed in /vendor/lib as sphal
libft2.so is removed from LLNDK, as it was LLNDK-private just because it is
referenced from VNDK-SPs, but it is no longer true because of VNDK
deprecation. This change adds libft2.so to have same sepolicy with other
sphal libraries, so it can be loaded from sphal libraries same as
before.

Bug: 326402813
Bug: 326338253
Test: N/A
Change-Id: Id8c1194da478bd4fc02e701230fd1a3c0b3c00be