Revert "Check added types/attributes on freeze test too"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/702f8dee-2391-4596-96d4-2eaa084fc600). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0331a4519308b9c34cea156cea27d96f273d7dcb