Revert "Restrict SDK sandbox unix_stream_socket connections"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e437441-f67d-4fc3-9073-33144e813990). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54e676950bfe206ca20bf45ae89fbb4a248a96e5