Revert "Restrict SDK sandbox unix_stream_socket connections"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78801072-cef9-4cb3-8276-0f9e3e45790f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2a754cd3778a5e0a9080405413d48fe56df0a53e