Revert "Restrict SDK sandbox unix_stream_socket connections"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/68204666-a226-4aac-82dc-51c92552957e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I16a234644f736ea542eefb862d2446ba7077a1ce