Revert "allow domain search for /metadata dir"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c7e2dfbb-404c-47a0-aead-cd1ccc1485b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib836dfd2229b3625390d12b4b276f56524f27a80