Revert "lmkd: Add sepolicy rules around bpf for lmkd"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f915fb46-b4b6-4990-8067-1cd4a51294d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I96a2cbf8ed7145f2b5796df0efb215fa8c1f1de5