Revert "lmkd: Add sepolicy rules around bpf for lmkd"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2381ccf2-5670-485c-93c0-22519874f7a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic28f736ccecd9f9d757df8f9b682187b90a1e3e9