Add ro.config.wallpaper property policies
ro.config.wallpaper is a property used by WallpaperManager to
determine the default wallpaper path. Basically, it's a mechanism to
override the wallpaper path from 'product' image (the other mechanism
is a combination of hardcoded '/product/wallpaper/image' +
ro.boot.hardware.color property).

ro.config.wallpaper should be defined in a build.prop file. But
because the property is not properly labeled (it is implicitly labeled
as u:object_r:config_prop:s0), it cannot be loaded by init process.

Test: m
Change-Id: Ic84ab2b9a21d8893b1934df260d1594eb0d07db4
Signed-off-by: Sergey Bobrenok <sibobrenok@salutedevices.com>