Revert "Add SELinux policy for storage areas"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1c6fec7e-b7bd-498b-a705-e736dc64328c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3a91284b5bfb60e2621385394d6411d4272946e