Revert "Thread: allow ot-rcp to bind a specific netif"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1d80bb9-8e0c-4e9b-bda0-131f6f1fbd58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 341213561
Change-Id: Ia5621dd5fa977a269533a67a3659c1b09c31b1b4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com