Revert "Add ro.fstype.data to indicate fs type of /data"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14e07f88-0bb2-4ffd-a9c1-465f617ae4cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343561082
Change-Id: I9cdcef46562974cbc1c0765fbb58f04cdfa2f4a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com