Revert "Add ro.fstype.data to indicate fs type of /data"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ac43f6d-5869-4f87-bbcd-4f193a0ee080). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343561216
Change-Id: Ibef2d1c758d12bb4204a374a13f57d5c22576ec2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com