Revert "Add ro.fstype.data to indicate fs type of /data"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/799bdb33-8893-43c9-93b7-788f88e1f1ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560933
Change-Id: I5b1f5b90104c4049425f8921a5a2fa573d6ffcce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com