Revert "Add ro.fstype.data to indicate fs type of /data"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e704d6d-a6ea-44b1-a143-9dff157a16d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343561125
Change-Id: Ic181459d0b7130bc3748b21fcd2a601aa658e1bc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com