Revert "Add ro.fstype.data to indicate fs type of /data"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03f3df84-5ca4-48b9-9fb7-3b4cda663aea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560867
Change-Id: I30ee73aa65deac7f4eda27e0c2b1cbab7b482850
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com