Revert "Add ro.fstype.data to indicate fs type of /data"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e7b0974-751a-45ed-8d0b-37c4244531e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560724
Change-Id: I7408a5c2ef951adaea0f002d50b8f7d21795221a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com