Revert "Add ro.fstype.data to indicate fs type of /data"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64e13983-6cff-4737-b49b-813e6298e430). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343560979
Change-Id: I03a3b9b687a79030a2bf005b3c8c8f14a003233c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com