Revert "Allow dexopt_chroot_setup to bind-mount dirs for increme..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3395f4a9-6f99-4073-aa4f-9c0ec639b25f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17dc7fc533e7d7abe9cad6f635c338577441888a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com