Revert "Introducing vm_tethering_service as system_server_service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0d76eb7-76a3-4bbd-9758-72095cdc602a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347148819
Change-Id: I078be3130dc25625866390973ee2f196c24cccfa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com