Revert "Introducing vm_tethering_service as system_server_service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3c6823e-6fbe-4a97-b810-1764e1b4f249). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 347407635
Change-Id: I9aabfdc463fe20060ec4290b4fef759822aaad82
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com