Revert "Add /system/bin/traced_relay to file_contexts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50dee470-a617-4c8f-bcad-04bbe490b250). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib071c337b2ec6c229e5e165160fee0230ee1a8f3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com