Revert "Add perfetto producer to servicemanager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1611620-a37b-4109-9078-18484377a207). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ab991c3f5911e10bd48c7046121cf85acd008fb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com