Revert "Add perfetto producer to servicemanager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c164a68d-8ff8-4fee-9923-9cdf1bf58742). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id672ed49112780d473810b1faa1828f3b032d782
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com