Revert "Add perfetto producer to servicemanager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59ff6b39-de1d-4e84-b303-39b0d8a67384). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If2317f54689fc7f705199fb84b3084c090be8db7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com