Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d1d77129-6e0a-43d9-a526-c723224b5a3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349242682
Change-Id: Ibc386c17034e4c219c0a1e62580c26d0e5673611
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com