Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d3a3c014-cf95-4744-b705-621739ea0217). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349242754
Change-Id: Id5efdf861af8bd4eed80db9297144994b60c5ccf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com