Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1b26595-f7da-419e-9fce-6001061fe8e3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349243034
Change-Id: I6c5ff90e8ff15d1d1e51892341a9b3027d54ae61
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com