Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3d3d7d97-78e7-4d6e-b283-6a39e7dd98da). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349242985
Change-Id: I7f4196d1ae8153e3d851b12260b0f8c38c31f0d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com