Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85c5e60a-9e60-4452-9ee9-9f5f900282e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349243264
Change-Id: I7a8795dc7bb9a352476f1e4328932824929f6082
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com