Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f3d5062-f753-4c40-8000-b2c6c217b754). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349242598
Change-Id: Ia239fc6f38a1e21d9bd5990430097e11b9bc0209
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com