Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7752affc-f08e-46aa-83d4-5c7061c7db76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349269115
Change-Id: I8631b9f5da9f45ad81f2efe839475885b614c59f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com