Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/171b08bf-fb95-4104-861b-30d42be44cb8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349243140
Change-Id: I451ea9d9479167898f0fae60a3c0587168dc20b8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com