Revert "Allow system server to access udc sysfs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ec91af68-af9b-469f-addd-439860d9d2ca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 349242875
Change-Id: I9c1242e7918511f9f77dbdadf0e106d65c05b762
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com