Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cdc6a82b-1740-42f0-ad10-3e1816cdbadf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 350001948
Change-Id: Ia869d64f9c769f9dc3238334e38c93aafed2db7c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com