Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9d98a45-e44f-4549-99ff-06c58e245f48). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 350001277
Change-Id: I9739ecae8c2042d8871737419ff93b49957e4ea7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com