Revert "Grant sigkill for dumpstate on su"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1883bf99-b8c6-4fef-a64f-ae39144e8a23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f51739e8cc5885b82acec053d95540df479aaf2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com