Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49b4e885-14ea-4a13-9f23-e57df71ed9d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib750367e78acf5e81b6a163922319031a1863a83
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com