Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/03443067-ace5-4450-b6ef-260e67a5e0ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e8d0bf82acf601821a855190174ac3c1e4669ff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com