Revert "Sepolicy for crosvm to play audio with aaudio"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/675dbfd5-bf20-4732-a864-369731712e82). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I084d276d9eabe2c9557ca968227fe0174adb319d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com