Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9225d52-8c88-4ac9-adc3-26c614abc560). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f794f4d3e360a5ed32f1d93568cd9f56186bae4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com