Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a16af4c-d59d-422d-822f-149223ac3b7d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79d793b0d63ccc0e833704d23a0eb511a9746245
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com