Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/96d9406e-50cb-434d-957a-9647c8049f0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcf2df09e00d62786f3f72b6fa1d0c69c556e8dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com