Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1a54b85-1485-47f7-8d3d-71af555c35f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibdbc96c4acdbc37e9e8ee04e543d65691fb00afc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com