Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e263976e-4bfd-4ab0-8e69-7a77a1e2f5c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e8c622a1d92ea29afde63619c4d31ff3064a36d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com