Revert "Fix the sepolicy issue around shell and oatdump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/371f336d-a10e-4b13-a0c2-3afc469cf734). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352184398
Change-Id: I55ee97c7f3fcc83044961dbf28ee87c1c94b6d8f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com