Revert "Fix the sepolicy issue around shell and oatdump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e31121bf-011e-4ed5-9d7f-e35cd5d093eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352184752
Change-Id: Id4b96b8e8193a5853edeee1a1d5b64eccfc9bc65
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com