Revert "Fix the sepolicy issue around shell and oatdump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/425cad47-5ac6-4675-86e4-d683f7aa71e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352184521
Change-Id: Ifb62276a2ff3ce6f5ca128e6c1ba2c6952ae7d47
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com