Revert "Fix the sepolicy issue around shell and oatdump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a76f63f6-fe1e-4b5c-8d1a-ae88dabb5a64). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352184202
Change-Id: I7d3713253cab91381e90611318958eba68727ac0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com