Revert "Fix the sepolicy issue around shell and oatdump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7d3a6f91-9660-4342-93db-34cf4665e980). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352184715
Change-Id: I7da2b08e42f045bf0fa847f12c8089bcf75e366a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com