Revert "Fix the sepolicy issue around shell and oatdump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f424de16-cd5c-4a76-82c8-68f702228a80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352184989
Change-Id: I194b541238ff5e564233e7fdb9e196c6a75b3d75
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com