Revert "Fix the sepolicy issue around shell and oatdump"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c70e4345-63b3-4f7e-a14d-75d90f1b0485). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 352184678
Change-Id: If6810a57ed4979ae77915746a0a77c4362e57bda
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com