Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/664b2044-be57-4b02-8c51-047aec22f68a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia318ba12c1729f165657d3fa5add99fcf0bce4e1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com