Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f218a7b-ff76-471e-abe0-1b4c761ea8cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353828307
Change-Id: Id0da659f31b17cd2b67ea7f2d2fbf5a5b35f82b0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com