Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f04f61a8-44e6-46fe-a13e-feee0b638830). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353828179
Change-Id: If765cf243e1abaf1e19f15a0be3765ffacb9fe3e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com