Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f17ddcee-3790-47b5-beaf-31fcada70591). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353828306
Change-Id: I0ce73a619f38f9a6a95e5de16945bb6412807cf7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com