Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c087545-2f38-4710-b38e-209c4f5cf112). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353828195
Change-Id: I90548d41f7a4b64447b8c884e472c419aa9eb6e2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com