Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/118da6c2-086e-43dd-95d9-48bf3bc76602). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353827925
Change-Id: I35ea05ae292d2f777675eb7aa18cbeb69db6e7b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com