Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed4e3f96-5e65-43b7-b6aa-fabdd4ded591). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353827701
Change-Id: Ia1a686fcbe0c2389c76d4f2fffeefbf0ffe8a48f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com