Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95b7a37d-9a55-4b65-9824-2b0c71773430). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353828087
Change-Id: Id3db7a35ae8c7e61087b4cec3a37f11bd39ebcd2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com