Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e414d88-e376-45a2-a4d6-4058dca15258). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353827700
Change-Id: I5934a13bf045b2bac441f49664172b30f695fe92
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com