Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d74addd-f049-4d13-a3f7-606f1b579fdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353828353
Change-Id: Iea0f7cf0cdc544ebe768e8a5851e54ca604ef8c4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com