Revert "Expose starting_at_board_api to access_vectors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9382a9cb-4576-4d4c-9dd3-b8d8b2dffea1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 353828003
Change-Id: I4aff71e75bc30f3eddbfb375ee3959330b0c37ed
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com