Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1c14bb5-ff38-495e-8a23-afc8404a8329). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I65b432b2d22ebe55a7e066e24a6f4e31532b037e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com