Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f2a4a71-ce72-4af6-a83d-24f6a9164d01). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia02d26a4b12d00db0da527eb632967826a68240b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com