Revert "sepolicy: remove all remaining qtaguid stuff."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f12999a0-3388-48f8-a460-dec20620385b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 357178435
Change-Id: I3006ec6ac0ca883023e5da10f7d6d956f4fc1d78
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com