Revert "Restrict vsock permissions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/206b0fdc-4e5a-4737-a5e8-d00c3fcae644). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa2fa65e7f01605eca20bed5694cb4f28735416b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com