Revert "Expose virtual_camera types"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/17f8a84c-0bb7-4d7c-88fc-10a7bd96873d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 357609913
Change-Id: I3bfa69f314852219793b32e9ddc3f2a149d2efc6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com