Erase scratch after OTA update
If the OTA update is installed successfully, there should be
the content of partitions from the installed OTA package.
But the cleaning of the scratch partition is missed now, so
scratch partition's content is applied on top of the OTA package.
This can lead to unexpected behaviour during the development
phase. This patch is intended to fix the described above issue.

Test: checked locally on my AOSP13-based device
Change-Id: I7718d8c8ac92190e2313f4789c603ae414c1d132
Signed-off-by: Anton Utkin <avutkin@sberdevices.ru>