Revert "Add microfuchsia selinux policies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1c0223f-4063-4c41-bafe-014da5351387). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361108146
Change-Id: I14b176d18ecd940ff6fcfa4c11caea90f2ee74f2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com