Allow app to kill crosvm
The process hierarchy is like this:

app -> virtmgr -> crosvm

Previously before this topic, when virtmgr is killed (e.g. for a bug, or
due to an attack) crosvm was reparented to init. crosvm was not reaped
by init because nobody actually didn't kill it. As a result, there's no
way for the app to kill the VM other then shutting down itself.

The other CL in this topic fixes that. This change adds two sepolicy
rules that the fix requires:

First, app is allowed to call getpgid on virtmgr (virtualizationservice)
so that it can get the process group id and call killpg on it.

Second, app is allowed to call sigkill on crosvm process. Note that this
doesn't mean that an app is allowed to kill random crosvm instances. The
app and crosvm have to have the same UID.

Bug: 280423229
Test: android/vm/vm_shell.sh start-microdroid --auto-connect
Open a new adb session and do the following:

1. su && kill -9 <pidof_virt_mgr>. vm is shutdown.

2. su media && kill -9 <pid_of_virtmgr>. "Operation not permitted" error
   is shown.

Change-Id: I424525e5d1c7fda314d2c3fa80dc46debf7c4791