Revert "Revert "Allow vendor init to access compaction_proactive..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f1ccc5b7-2d53-423a-a7e2-591c74a0a6cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363306995
Change-Id: I30a83cd8d788e5337eacdcd6ef565ed4d66fbc3c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com