Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f52047b3-0bc0-4b08-be8a-120d062fc677). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363980706
Change-Id: Ia30ae8460dc4d85f986ce8502163b8184e3c6602
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com