Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fcc3a2fc-e71f-4080-98d3-ea79e5ff4a26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363980608
Change-Id: I60761298d317ffc24b9001bcb078615c2f6d0192
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com