Revert "API flagging macro until/starting_at for private and public"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/835aa2bd-6a83-46ef-b50a-0619b5314161). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363981114
Change-Id: I8347a1b9f8c82f28ae1c55b74685471418e1dfbb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com